add step 16
the feature test in the walkthrough was not correct, the before is
done without any feature keyword above.  Also they say to add the
test into the viewing_link file, I did that putting together all
feature tests but in this cases the file should be name something
different.

Also, in the walkthrough in the app.rb they add the following line
tag = Tag.first(name: params[:name])
that only gives you the first link with the tag.  We want all the
tags with the name, ie. the line needs to be:

tag = Tag.all(name: params[:name])

Finally, the line:
@links = tag ? tag.links : []

looks a bit strange but it is ok, we check if tag is nil on the
assignmen, if so, we assign @links to the empty array, if not we
assign @links the tag.links value.