step 21
added all the relevant items, and they worked.  I did not know how
to add the null constraints in the database, and as usual the
workthrough is not correct, the right way is using the following in
psql:
\c bookmark_manager_development;
\dt  (to show tables)
DROP TABLE users;
CREATE TABLE "users" ("id" SERIAL NOT NULL, "email" VARCHAR(50) NOT NULL,
"password_digest" VARCHAR(60), PRIMARY KEY("id"));
(again missing semicolon in walkthrough)

Regarding validation of email addresses in HTML5, it is a very weak feature,
emails such as ada@mia are accepted as valid email addresses.  The required
in the email form works better, however it does not make the feature test pass,
so they must be ways around.