step 22
definition of test was easy, similar to previous ones.  Validation
was easy too, just add unique triger in email i.e.
property :email, String, required: true, format: :email_address, unique: true
adding the error was a bit more tricky, as we already had the
mismatching password and confirmation.  But exploring the error method was
the right way forward.
Again the workthrough continues to confuse  with the flash, now the change the argument
from :notice to :error and continues to be wrong.